+++ packages/bottom-sheet/src/component.screenshots.test.tsx
@@ -4,6 +4,7 @@ import {
     openBrowserPage,
     matchHtml,
     closeBrowser,
+    Knobs,
 } from '../../screenshot-utils';
 
 const screenshotTesting = setupScreenshotTesting({
@@ -129,4 +130,38 @@ describe('BottomSheet | interactions tests', () => {
             await closeBrowser({ browser, context, page });
         }
     });
+});
+
+const magneticAreaTest = async (knobs: Knobs) => {
+    const pageUrl = createStorybookUrl({
+        componentName: 'BottomSheet',
+        testStory: true,
+        knobs: {
+            open: true,
+            initialHeight: 'full',
+            magneticAreas: '["10%25","50%25","100%25"]',
+            children: 'Magnetic areas test',
+            ...knobs,
+        },
+    });
+
+    const { browser, context, page, css } = await openBrowserPage(pageUrl);
+
+    try {
+        await matchHtml({ context, page, expect, css, screenshotOpts: { fullPage: true } });
+    } catch (error) {
+        // eslint-disable-next-line no-console
+        console.error((error as Error).message);
+
+        throw error;
+    } finally {
+        await closeBrowser({ browser, context, page });
+    }
+};
+
+describe('BottomSheet | magnetic areas test', () => {
+    test('10%', () => magneticAreaTest({ initialActiveAreaIndex: 0 }));
+    test('50%', () => magneticAreaTest({ initialActiveAreaIndex: 1 }));
+    test('100%', () => magneticAreaTest({ initialActiveAreaIndex: 2 }));
+    test('negative 20%', () => magneticAreaTest({ magneticAreas: '[0,"-20%25"]' }));
 });