+++ packages/select/src/components/option/Component.tsx
@@ -2,8 +2,8 @@ import React, { FC, isValidElement } from 'react';
 import cn from 'classnames';
 
 import { OptionProps } from '../../typings';
-import { Checkmark as DefaultMobileCheckmark } from '../base-select-mobile/checkmark';
 import { Checkmark as DefaultCheckMark } from '../checkmark';
+import { Checkmark as DefaultMobileCheckmark } from '../checkmark-mobile';
 
 import styles from './index.module.css';