+++ packages/select/src/components/select-mobile/Component.tsx
@@ -2,12 +2,16 @@ import React, { forwardRef, ReactNode } from 'react';
 
 import { BottomSheetProps } from '@alfalab/core-components-bottom-sheet';
 
-import { BaseSelectProps } from '../../typings';
+import { useSelectWithApply, UseSelectWithApplyProps } from '../../presets/useSelectWithApply/hook';
+import { AnyObject, BaseSelectProps } from '../../typings';
 import { Arrow as DefaultArrow } from '../arrow';
 import { BaseSelectMobile } from '../base-select-mobile';
+import { Footer } from '../base-select-mobile/footer';
 import { Field as DefaultField } from '../field';
 import { Optgroup as DefaultOptgroup } from '../optgroup';
 import { Option as DefaultOption } from '../option';
+import { OptionsList as DefaultOptionsList } from '../options-list';
+import { VirtualOptionsList as DefaultVirtualOptionsList } from '../virtual-options-list';
 
 export type AdditionalMobileProps = {
     /**
@@ -26,10 +30,22 @@ export type AdditionalMobileProps = {
      * Дополнительные пропсы шторки
      */
     bottomSheetProps?: Partial<BottomSheetProps>;
+
+    /**
+     * Показывать кнопку очистки
+     */
+    showClear?: UseSelectWithApplyProps['showClear'];
+
+    /**
+     * Показывать пункт "Выбрать все"
+     */
+    showSelectAll?: UseSelectWithApplyProps['showSelectAll'];
 };
 
 export type SelectMobileProps = Omit<BaseSelectProps, 'Checkmark' | 'onScroll'> &
     AdditionalMobileProps;
+
+const VIRTUAL_OPTIONS_LIST_THRESHOLD = 30;
 
 export const SelectMobile = forwardRef(
     (
@@ -46,34 +62,66 @@ export const SelectMobile = forwardRef(
             optionsSize = 'm',
             fieldProps = {},
             optionProps = {},
+            optionsListProps = {},
             Arrow = DefaultArrow,
             Field = DefaultField,
             Optgroup = DefaultOptgroup,
             Option = DefaultOption,
+            selected,
+            options,
+            OptionsList = options.length > VIRTUAL_OPTIONS_LIST_THRESHOLD
+                ? DefaultVirtualOptionsList
+                : DefaultOptionsList,
+            onChange,
+            bottomSheetProps,
+            showClear = true,
+            showSelectAll,
             ...restProps
         }: SelectMobileProps,
         ref,
-    ) => (
-        <BaseSelectMobile
-            ref={ref}
-            autocomplete={autocomplete}
-            multiple={multiple}
-            allowUnselect={allowUnselect}
-            disabled={disabled}
-            closeOnSelect={closeOnSelect}
-            circularNavigation={circularNavigation}
-            defaultOpen={defaultOpen}
-            open={openProp}
-            size={size}
-            optionsSize={optionsSize}
-            fieldProps={fieldProps}
-            optionProps={optionProps}
-            Arrow={Arrow}
-            Field={Field}
-            Optgroup={Optgroup}
-            Option={Option}
-            isBottomSheet={true}
-            {...restProps}
-        />
-    ),
+    ) => {
+        const applyProps = useSelectWithApply({
+            optionsListProps: {
+                ...(optionsListProps as AnyObject),
+                Footer,
+            },
+            OptionsList,
+            selected,
+            options,
+            onChange,
+            showClear,
+            showSelectAll,
+        });
+
+        return (
+            <BaseSelectMobile
+                ref={ref}
+                autocomplete={autocomplete}
+                multiple={multiple}
+                allowUnselect={allowUnselect}
+                disabled={disabled}
+                closeOnSelect={closeOnSelect}
+                circularNavigation={circularNavigation}
+                defaultOpen={defaultOpen}
+                open={openProp}
+                size={size}
+                optionsSize={optionsSize}
+                fieldProps={fieldProps}
+                optionProps={optionProps}
+                Arrow={Arrow}
+                Field={Field}
+                Optgroup={Optgroup}
+                Option={Option}
+                isBottomSheet={true}
+                options={options}
+                selected={selected}
+                onChange={onChange}
+                OptionsList={OptionsList}
+                bottomSheetProps={bottomSheetProps}
+                {...(optionsListProps as AnyObject)}
+                {...restProps}
+                {...(multiple && !bottomSheetProps?.actionButton && applyProps)}
+            />
+        );
+    },
 );