+++ packages/bottom-sheet/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log
 
+## 6.2.0
+
+### Minor Changes
+
+### [#615](https://github.com/core-ds/core-components/pull/615)
+
+-   Добавлены магнитные области. Полное описание изменений в [#615](https://github.com/core-ds/core-components/pull/615)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   backdrop@3.0.6
+    -   navigation-bar@0.2.0
+    -   base-modal@5.1.3
+
+## 6.1.0
+
+### Minor Changes
+
+### [#604](https://github.com/core-ds/core-components/pull/604)
+
+-   Добавлен новый prop backgroundColor
+
 ## 6.0.5
 
 ### Patch Changes