+++ packages/calendar-with-skeleton/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 5.0.38
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@6.2.17
+
+## 5.0.37
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@6.2.16
+
 ## 5.0.36
 
 ### Patch Changes