+++ packages/select-with-tags/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 7.0.28
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@14.0.0
+
+## 7.0.27
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@13.2.20
+
 ## 7.0.26
 
 ### Patch Changes