+++ packages/side-panel/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 3.0.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   navigation-bar@0.2.0
+    -   base-modal@5.1.3
+    -   drawer@4.2.10
+
 ## 3.0.5
 
 ### Patch Changes