+++ packages/slider-input/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 7.1.13
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   slider@4.2.0
+
+## 7.1.12
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   slider@4.1.6
+
 ## 7.1.11
 
 ### Patch Changes