+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 7.3.16
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   picker-button@10.3.16
+
+## 7.3.15
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   picker-button@10.3.15
+
 ## 7.3.14
 
 ### Patch Changes